osconfg agent: Only take package action when needed
