+++ go/packages/googet_test.go
@@ -1,0 +1,48 @@
+//  Copyright 2019 Google Inc. All Rights Reserved.
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+
+package packages
+
+import (
+	"errors"
+	"testing"
+)
+
+func TestInstallGooGetPackages(t *testing.T) {
+	run = getMockRun([]byte("TestInstallGooGetPackages"), nil)
+	if err := InstallGooGetPackages(pkgs); err != nil {
+		t.Errorf("unexpected error: %v", err)
+	}
+}
+
+func TestInstallGooGetPackagesReturnsError(t *testing.T) {
+	run = getMockRun([]byte("TestInstallGooGetPackagesReturnsError"), errors.New("Could not install package"))
+	if err := InstallGooGetPackages(pkgs); err == nil {
+		t.Errorf("did not get expected error")
+	}
+}
+
+func TestRemoveGooGet(t *testing.T) {
+	run = getMockRun([]byte("TestRemoveGooGet"), nil)
+	if err := RemoveGooGetPackages(pkgs); err != nil {
+		t.Errorf("unexpected error: %v", err)
+	}
+}
+
+func TestRemoveGooGetReturnError(t *testing.T) {
+	run = getMockRun([]byte("TestRemoveGooGetReturnError"), errors.New("Could not find package"))
+	if err := RemoveGooGetPackages(pkgs); err == nil {
+		t.Errorf("did not get expected error")
+	}
+}
