+++ go/packages/yum.go
@@ -1,0 +1,122 @@
+//  Copyright 2019 Google Inc. All Rights Reserved.
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+
+package packages
+
+import (
+	"fmt"
+	"os/exec"
+	"runtime"
+	"strings"
+	"syscall"
+
+	"github.com/GoogleCloudPlatform/compute-image-tools/go/osinfo"
+)
+
+var (
+	yum string
+
+	yumInstallArgs     = []string{"install", "-y"}
+	yumRemoveArgs      = []string{"remove", "-y"}
+	yumUpdateArgs      = []string{"update", "-y"}
+	yumCheckUpdateArgs = []string{"check-update", "--quiet"}
+)
+
+func init() {
+	if runtime.GOOS != "windows" {
+		yum = "/usr/bin/yum"
+	}
+}
+
+// InstallYumPackages installs yum packages.
+func InstallYumPackages(pkgs []string) error {
+	args := append(yumInstallArgs, pkgs...)
+	out, err := run(exec.Command(yum, args...))
+	if err != nil {
+		return err
+	}
+	var msg string
+	for _, s := range strings.Split(string(out), "\n") {
+		msg += fmt.Sprintf(" %s\n", s)
+	}
+	DebugLogger.Printf("yum install output:\n%s\n", msg)
+	return nil
+}
+
+// RemoveYumPackages removes yum packages.
+func RemoveYumPackages(pkgs []string) error {
+	args := append(yumRemoveArgs, pkgs...)
+	out, err := run(exec.Command(yum, args...))
+	if err != nil {
+		return err
+	}
+	var msg string
+	for _, s := range strings.Split(string(out), "\n") {
+		msg += fmt.Sprintf(" %s\n", s)
+	}
+	DebugLogger.Printf("yum remove output:\n%s\n", msg)
+	return nil
+}
+
+// update yum packages
+func yumUpdate(run runFunc) error {
+	if _, err := run(exec.Command(yum, yumUpdateArgs...)); err != nil {
+		return err
+	}
+	return nil
+}
+
+func yumUpdates() ([]PkgInfo, error) {
+	out, err := exec.Command(yum, yumCheckUpdateArgs...).CombinedOutput()
+	// Exit code 0 means no updates, 100 means there are updates.
+	if err == nil {
+		return nil, nil
+	}
+	if exitErr, ok := err.(*exec.ExitError); ok {
+		if status, ok := exitErr.Sys().(syscall.WaitStatus); ok && status.ExitStatus() == 100 {
+			err = nil
+		}
+	}
+	if err != nil {
+		return nil, fmt.Errorf("error checking yum upgradable packages: %v, stdout: %s", err, out)
+	}
+	/*
+
+	   foo.noarch 2.0.0-1 repo
+	   bar.x86_64 2.0.0-1 repo
+	   ...
+	   Obsoleting Packages
+	   ...
+	*/
+
+	lines := strings.Split(strings.TrimSpace(string(out)), "\n")
+	if len(lines) == 0 {
+		return nil, nil
+	}
+
+	var pkgs []PkgInfo
+	for _, ln := range lines[1:] {
+		pkg := strings.Fields(ln)
+		if len(pkg) == 2 && pkg[0] == "Obsoleting" && pkg[1] == "Packages" {
+			break
+		}
+		if len(pkg) != 3 {
+			DebugLogger.Printf("%s does not represent a yum update\n", ln)
+			continue
+		}
+		name := strings.Split(pkg[0], ".")
+		pkgs = append(pkgs, PkgInfo{Name: name[0], Arch: osinfo.Architecture(name[1]), Version: pkg[1]})
+	}
+	return pkgs, nil
+}
