+++ go/packages/yum_test.go
@@ -1,0 +1,48 @@
+//  Copyright 2019 Google Inc. All Rights Reserved.
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+
+package packages
+
+import (
+	"errors"
+	"testing"
+)
+
+func TestInstallYumPackages(t *testing.T) {
+	run = getMockRun([]byte("TestInstallYumPackages"), nil)
+	if err := InstallYumPackages(pkgs); err != nil {
+		t.Errorf("unexpected error: %v", err)
+	}
+}
+
+func TestInstallYumPackagesReturnsError(t *testing.T) {
+	run = getMockRun([]byte("TestInstallYumPackagesReturnsError"), errors.New("Could not install package"))
+	if err := InstallYumPackages(pkgs); err == nil {
+		t.Errorf("did not get expected error")
+	}
+}
+
+func TestRemoveYum(t *testing.T) {
+	run = getMockRun([]byte("TestRemoveYum"), nil)
+	if err := RemoveYumPackages(pkgs); err != nil {
+		t.Errorf("unexpected error: %v", err)
+	}
+}
+
+func TestRemoveYumReturnError(t *testing.T) {
+	run = getMockRun([]byte("TestRemoveYumReturnError"), errors.New("Could not find package"))
+	if err := RemoveYumPackages(pkgs); err == nil {
+		t.Errorf("did not get expected error")
+	}
+}
